requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.24.3
tensorflow==2.14.0
requests==2.31.0
pydantic==2.5.2
openai==1.3.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
tqdm==4.66.1
httpx==0.25.2 